requirements.txt
aiogram
web3
solana
solders
google-generativeai
asyncpg
python-dotenv
aiohttp
loguru
sqlalchemy
psycopg2-binary
tenacity
